backend/requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
passlib==1.7.4
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
